% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{add_location_prefix}
\alias{add_location_prefix}
\title{Add location prefix}
\usage{
add_location_prefix(cameras, locations, logger = NA)
}
\arguments{
\item{cameras}{a character vector of camera names}

\item{locations}{a character vector of locations (must be the same length as \code{cameras})}

\item{logger}{a \code{log4r} \code{logger} object if you want logging (can be created with \code{create_logger}),
else \code{NA}.}
}
\value{
A character vector of the same length as \code{cameras} where each value is prefixed
with the corresponding value of \code{locations} (in the same order).
}
\description{
Adds a prefix to a set of cameras, previously verifying that the prefix is not in the camera
name.
}
\examples{
cameras <- c("A01", "B01", "KAR_A01")
locations <- c("KAR", "KAR", "KAR")
add_location_prefix(cameras, locations)
}
