% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{check_standard_colnames}
\alias{check_standard_colnames}
\title{Check standard column names}
\usage{
check_standard_colnames(
  standard_colnames,
  classifier = c("zooniverse", "traptagger", "digikam")
)
}
\arguments{
\item{standard_colnames}{The dataframe to check}

\item{classifier}{The classifier.}
}
\value{
Nothing or stops execution if standard_colnames is not in
compliance with expectations.
}
\description{
Check that the standard column names match expectations, i.e.
is a dataframe with >= 2 columns named like the classifier object and 'new'.
}
