% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_camera_location_df}
\alias{clean_camera_location_df}
\title{Clean locations and cameras}
\usage{
clean_camera_location_df(df)
}
\arguments{
\item{df}{a dataframe that must have columns \code{cameraID}, \code{locationID} and \code{classifier}.}
}
\value{
The dataframe with cleaned columns \code{cameraID}, \code{locationID} and \code{eventID}.
}
\description{
Cleans locations and cameras for a dataframe
}
\details{
For column \code{locationID}:
\itemize{
\item The location code \code{DHP} is replaced with \code{OVE} if the corresponding camera code starts with 'O'.
\item The location code \code{KGA} is replaced with \code{KHO} if the corresponding camera code starts with 'KHO'.
}

For column \code{cameraID}:
\itemize{
\item For TrapTagger data: will remove the leading location code part for all data
(eg if \code{location} is \code{ATH}, will change \code{cameras} \code{ATH_A01} -> \code{A01}).
Also, if the location code is \code{KHO}, \code{SAM} or \code{TSW}: will remove the dash in
the camera name (e.g \code{KHO_E_A01} -> \code{EA01})
\item For Zooniverse data: if the location code is \code{KHO}, will replace
\code{KHOG} with \code{E} and  \code{KHOL} with \code{M} in \code{cameras}.
If the location code is \code{DHP}, will remove leading \code{D} in \code{cameraID}.
If the location code is \code{OVE}, will remove leading \code{O} in \code{cameraID}.
\item For column \code{eventID}: the event ID formatted as season#cam_site#roll#event_no.
}
}
