% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_cameras}
\alias{clean_cameras}
\title{Clean cameras}
\usage{
clean_cameras(
  cameras,
  locations,
  classifiers,
  silence_warnings = FALSE,
  logger = NA
)
}
\arguments{
\item{cameras}{Cameras vector}

\item{locations}{locations vector}

\item{classifiers}{Classifier vector. If the classifier is \code{NULL},
the function will apply all Zooniverse corrections.}

\item{silence_warnings}{Should a warning be displayed when \code{locationID} is \code{NA}?}

\item{logger}{a \code{log4r} \code{logger} object if you want logging (can be created with \code{create_logger}),
else \code{NA}.}
}
\value{
a character vector of cameras with the same length as input.
}
\description{
Standardize cameras names
}
\details{
For all data: will add the location prefix.

For TrapTagger data:
\itemize{
\item if the location code is \code{KHO}: will remove the dash in the camera name (e.g \code{KHO_E_A01} -> \code{KHO_EA01})
\item if the location code is \code{SAM}: will remove the dash in the camera name (e.g \code{SAM_B_A01} -> \code{SAM_BA01})
\item if the location code is \code{SAM}: will remove the dash in the camera name (e.g \code{TSW_L_A01} -> \code{TSW_LA01})
}

For Zooniverse data:
\itemize{
\item if the location code is \code{KHO}, will replace \code{KHOG} with \code{E} and  \code{KHOL} with \code{M} in \code{cameras} (and add location prefix).
\item if the location code is \code{DHP}, will remove leading \code{D} in \code{cameras}  (and add location prefix).
\item if the location code is \code{OVE}, will remove leading \code{O} in \code{cameras}  (and add location prefix).
}
}
\examples{
cameras <- c("KHO_E_A01", "OCO2")
locations <- c("KHO", "OVE")
classifiers <- c("traptagger", "zooniverse")
clean_cameras(cameras, locations, classifiers)
}
