% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean_locations}
\alias{clean_locations}
\title{Clean location}
\usage{
clean_locations(cameras, locations, logger = NA)
}
\arguments{
\item{cameras}{characgter vector of cameras}

\item{locations}{character vector of locations (same length)}

\item{logger}{a \code{log4r} \code{logger} object if you want logging (can be created with \code{create_logger}),
else \code{NA}.}
}
\value{
a character vector of locations with the same length as input.
}
\description{
Corrects potentially incorrect locations based on the content of cameras.
Indeed, for Zooniverse data, cameras in OVE have the locationID DHP and
cameras in KHO have the locationID KGA.
}
\details{
The code \code{DHP} is replaced with \code{OVE} if the corresponding camera code starts with 'O'.
The code \code{KGA} is replaced with \code{KHO} if the corresponding camera code starts with 'KHO'.
}
\examples{
clean_locations(c("OA01", "KHOGA01"), c("DHP", "KGA"))
}
