% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_handling.R
\name{create_logger}
\alias{create_logger}
\title{Create logger}
\usage{
create_logger(my_logfile)
}
\arguments{
\item{my_logfile}{path to the log file (character).}
}
\value{
Returns a logger and creates a logfile at the given path. If the path given does not
exist, also creates this path.
}
\description{
Initializes a logger with a given logfile.
}
