% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{digikam}
\alias{digikam}
\title{Digikam sample data}
\format{
A data frame with 100 rows and 22 variables:
\describe{
\item{\code{X}}{integer Row names (read by R when reading the csv file)}
\item{\code{Station}}{character Camera}
\item{\code{Species}}{character Species}
\item{\code{DateTimeOriginal}}{character Date and time}
\item{\code{Date}}{character Date}
\item{\code{Time}}{character Time}
\item{\code{delta.time.secs}}{integer Time elapsed since this species was last seen at this camera (seconds). Not relevant because rows were permuted.}
\item{\code{delta.time.mins}}{double Time elapsed since this species was last seen at this camera (minutes). Not relevant because rows were permuted.}
\item{\code{delta.time.hours}}{double Time elapsed since this species was last seen at this camera (hours). Not relevant because rows were permuted.}
\item{\code{delta.time.days}}{double Time elapsed since this species was last seen at this camera (days). Not relevant because rows were permuted.}
\item{\code{Directory}}{character Local directory where the original photo is}
\item{\code{FileName}}{character Name of the original photo on the local storage}
\item{\code{EXIF.Model}}{character Exif info}
\item{\code{EXIF.Make}}{character Exif info}
\item{\code{metadata_Species}}{character Species (other column)}
\item{\code{metadata_Number}}{character Species count}
\item{\code{metadata_Behaviour}}{character Tagged behaviors}
\item{\code{metadata_Sex}}{character Tagged sex}
\item{\code{n_images}}{integer Number of pictures associated to this event?}
\item{\code{metadata_young_present}}{character Tagged youngs (Yes/No)}
\item{\code{metadata_Numberofindividuals}}{logical Tagged number of individuals on the picture}
\item{\code{HierarchicalSubject}}{character Summary column for all metadata_...}
}
}
\usage{
digikam
}
\description{
A dataset mimicking typical Digikam data (randomized rows)
}
\keyword{datasets}
