% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{fill_capture_info}
\alias{fill_capture_info}
\title{Fill capture info}
\usage{
fill_capture_info(df, classifier = c("zooniverse", "traptagger", "digikam"))
}
\arguments{
\item{df}{a dataframe for which columns should be filled.}

\item{classifier}{The classifier}

\item{remove_temp}{should the column 'eventID' be removed if present?}
}
\value{
A dataframe with values filled for locationID, cameraID,
roll and capture. Also fills season if Zooniverse.
}
\description{
Fills different columns related to capture info.
}
