% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{fill_capture_info_digikam}
\alias{fill_capture_info_digikam}
\title{Fill capture info for Digikam}
\usage{
fill_capture_info_digikam(df)
}
\arguments{
\item{df}{The dataframe to fill. Must have a column 'locationID'}
}
\value{
The dataframe with completed values for columns 
'season', 'roll' and 'captureID'. season' defaults to 1; 
'roll' is extracted from 'file_path_1'; 
'captureID' is the i-th picture for a given camera.
}
\description{
Fills some data on capture event for Digikam-format data.
Fills season, roll and capture (as character).
}
