% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{fill_capture_info_traptagger}
\alias{fill_capture_info_traptagger}
\title{Fill capture info TrapTagger}
\usage{
fill_capture_info_traptagger(df)
}
\arguments{
\item{df}{The dataframe to complete}
}
\value{
The dataframe with values completed for the captureID
info columns. If no info, it remains NA.
}
\description{
Completes capture info for TrapTagger from the eventID.
Completes values for columns cameraID, roll, captureID, location and 
adds column season (NA). These columns are character.
}
