% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{fncols}
\alias{fncols}
\title{Add columns that don't exist}
\usage{
fncols(data, cname, approx = FALSE)
}
\arguments{
\item{data}{the dataframe to add columns to}

\item{cname}{a character vector with names that should be added (these names can be present in the data already)}
}
\value{
The dataframe with all original columns + those that are in 'cname'
}
\description{
Helper function to add columns that are not in the original data (code found on https://stackoverflow.com/questions/45857787/adding-column-if-it-does-not-exist)
}
