% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{get_cameraID}
\alias{get_cameraID}
\title{Get camera ID}
\usage{
get_cameraID(locationID, camera, classifier)
}
\arguments{
\item{locationID}{the location code vector (usually 3 letters)}

\item{camera}{the "old" camera code vector, which is expected not to contain the locationID already.
In case it does, then it will not be added and give a message.}

\item{classifier}{The classifier used. If it is traptagger, will not display message
because it is expected that cameraID will already be locationID_camera.
It is optional (if not specified, it will display the message by default)}
}
\value{
a vector of same lengths as locationID and camera with pasted
locationID_camera. If camera is already in format locationID_camera, 
then it does not changes and displays a message.
}
\description{
Returns the complete cameraID as locationID_camera.
}
