% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_handling.R
\name{get_csv_files_and_folders}
\alias{get_csv_files_and_folders}
\title{Get csv files and folder}
\usage{
get_csv_files_and_folders(input, except, basepath)
}
\arguments{
\item{input}{a character vector of valid paths: can be files or folders, or a mix of both}

\item{except}{files to ignore (optional): the path might be an absolute path
or the relative path from \code{basepath}.}

\item{basepath}{the part of the path that should be ignored when copying final
files (i.e. absolute path inside one's comupter that should not be copied in final file.)}
}
\value{
A dataframe with columns \code{folders} and \code{files}, where \code{folders} are the
paths  up to a given file, and \code{files} are the files paths from \code{folders}.
}
\description{
Get all csv files from the input character vector.
\code{input} can be a vector of mixed files and folders.
If an element of \code{input} is a folder, the function will list
all files within \code{input}; if he element is a file,
the function will only list this file.
}
\examples{
\dontrun{
get_csv_files_and_folders(input = c("path/to/datafolder/KGA",
                                    "path/to/datafolder/ATH_Roll1_Snapshot.csv"), 
                                    except = "KGA/KGA-KHO_together/*", 
                                    basepath = "path/to/datafolder")
# Using absolute path for except is alsp valid                                 
get_csv_files_and_folders(input = c("path/to/datafolder/KGA",
                                    "path/to/datafolder/ATH_Roll1_Snapshot.csv"), 
                          except = "path/to/datafolder/KGA/KGA-KHO_together/*", 
                          basepath = "path/to/datafolder")
}
}
