% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{get_eventID}
\alias{get_eventID}
\title{Get eventID}
\usage{
get_eventID(locationID, cameraID, roll, captureID)
}
\arguments{
\item{locationID}{character vector of location}

\item{cameraID}{character vector of cam_site}

\item{roll}{character vector of roll}

\item{captureID}{character vector of captureID}
}
\value{
A character vector of the same length of the inputs (or the
longest input) with fields formatted as 
season#cam_site#roll#event_no
}
\description{
Get the eventID from the capture info. All vectors must be the
same length or they will be recycled.
}
