% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_handling.R
\name{get_files_and_folder}
\alias{get_files_and_folder}
\title{Get files and folder}
\usage{
get_files_and_folder(input, except)
}
\arguments{
\item{input}{a character vector of valid paths: can be files or folders}

\item{except}{files to ignore (optional)}
}
\value{
A dataframe with columns folder and files, where folder is the 
folder up to a given file and file is the file.
}
\description{
Get all files from the input character vector. Either the basename from input
if they are files, or all files into input recursively if input are folders.
}
