% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_handling.R
\name{get_final_filename}
\alias{get_final_filename}
\title{Get final filemane}
\usage{
get_final_filename(df)
}
\arguments{
\item{df}{The dataframe to be copied. Must have columns \code{locationID}, \code{season}, \code{roll}.}
}
\value{
The filename for this file in the format \code{locationID_Sseason_Rroll.csv}
It there are several values in \code{locationID}, \code{season} or \code{roll}, they are separated by a dash
in the filename: \code{locationID1-locationID2}...
}
\description{
Return the filename from the file columns.
}
\examples{
zooniverse_std <- standardize_snapshot_df(df = zooniverse, standard = standard)
get_final_filename(zooniverse_std)
}
