% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_handling.R
\name{get_logfile_name}
\alias{get_logfile_name}
\title{Name logfile}
\usage{
get_logfile_name()
}
\value{
A string with format "log__YYYY-MM-DD_HH:MM:SS.log"
where YYYY-MM-DD_HH:MM:SS is the current date/time.
}
\description{
Create a name for the logfile from an input file/folder.
}
