% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{get_NAs}
\alias{get_NAs}
\title{Get NAs}
\usage{
get_NAs(transformed_data, origin_data, nmax = 50, collapse = TRUE)
}
\arguments{
\item{transformed_data}{The data that was transformed which we want to check for NAs}

\item{origin_data}{The original, untransformed data}

\item{nmax}{Maximum number of values to return. If zero, returns all values}

\item{collapse}{Whether to collapse the values into a character (collapses 
with ', ')}
}
\value{
Returns a vector of the original values that were transformed 
to NAs (if collapse is FALSE). If collapse is TRUE, returns a character
containing the original values that were transformed to NA.
If no NAs were found, returns a vector numeric(0)
}
\description{
Get the original data that are NA in a new vector.
}
