% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_handling.R
\name{get_relative_path}
\alias{get_relative_path}
\title{Get relative path}
\usage{
get_relative_path(f, wd)
}
\arguments{
\item{f}{the full path to the folder/file.}

\item{wd}{a string describing the part of the path to delete.}
}
\value{
A string with the truncated path, without leading/terminal "/"
if f is a folder.
}
\description{
Get the relative path to a file or folder.
}
