% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{guess_classifier}
\alias{guess_classifier}
\title{Guess classifier}
\usage{
guess_classifier(colnames_df)
}
\arguments{
\item{colnames_df}{A character vector of column names.}
}
\value{
The classifier: either zooniverse, digikam or traptagger.
}
\description{
Guesses the classifier used to annotate the data based on the column names
given in colnames_df.
}
\examples{
guess_classifier(colnames(zooniverse))
guess_classifier(colnames(traptagger))
guess_classifier(colnames(digikam))
}
