% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{keep_only_expected_columns}
\alias{keep_only_expected_columns}
\title{Keep only expected columns}
\usage{
keep_only_expected_columns(df, expected, verbose, approx)
}
\arguments{
\item{df}{The dataframe to filter}

\item{expected}{Character vector of expected column names.}

\item{verbose}{Should a message be displayed if some columns are deleted?}

\item{approx}{Should partial name matching be used to match expected columns?}
}
\value{
The dataframe with all columns in expected, and no others.
If a column was not originally there, then it is filled with NAs.
}
\description{
Keep only expected columns
}
