% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_handling.R
\name{list_csv_in_folder}
\alias{list_csv_in_folder}
\title{List all files in folder}
\usage{
list_csv_in_folder(folder, except)
}
\arguments{
\item{folder}{the folder to list files in}

\item{except}{Paths to ignore (optional): these are the relative
path from 'folder'. They are either full file names, or a path
with partial matching like "folder/*".}
}
\value{
A character vector with relative paths of all files in the folder 'folder',
except the files that are indicated by 'except'.
}
\description{
This function lists all csv files in a base folder. It works recursively 
(lists all files in the folder and all files in subfolders of this folder)
}
