% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeSnapshot-package.R
\docType{package}
\name{PEcAn.standardizeSnapshot}
\alias{PEcAn.standardizeSnapshot}
\title{standardizeSnapshot is a R package to standardize camera trap records files from the Snapshot Safari project.}
\description{
Snapshot Safari data comes in different standards, 
following the method that was used to classify pictures 
(Zooniverse, TrapTagger or Digikam). This package allows to standardize 
all data sources to a unique file format and then cleans the files to 
homogenize records.
See Snapshot Safari: a large-scale collaborative to monitor Africa’s remarkable biodiversity (Pardo et al, 2021)
(https://www.sajs.co.za/article/view/8134)
}
