% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{prepare_digikam}
\alias{prepare_digikam}
\title{Prepare Digikam data}
\usage{
prepare_digikam(df)
}
\arguments{
\item{df}{Dataframe to standardize (must have columns 'metadata_Behaviour' and 'metadata_young_present').}
}
\value{
Returns the dataframe without 'X' column, 'Directory' and 'Filename' merged
into 'file_path_1' column, 'metadata_Behaviour' splitted into up to 6 columns
'Eating', 'Drinking', 'Resting', 'Standing', 'Moving', 'Interacting' (depending on the
behavoirs that were present in the data) and the 'metadata_young_present' column recoded
so that 'Yes'/'No' becomes 0/1.
}
\description{
Standardizes Digikam data to be ready to be standardized: removes 'X' index column,
(if it exists), merges 'Directory' and 'Filename' to create a path, 
splits 'metadata_Behaviour' column into the relevant behaviors and 
recodes 'metadata_young_present' into 0 and 1.
}
