% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{prepare_traptagger}
\alias{prepare_traptagger}
\title{Prepare TrapTagger data}
\usage{
prepare_traptagger(df)
}
\arguments{
\item{df}{The dataframe to standardize (must have columns starting with capture_label
and column timestamp)}
}
\value{
Returns the dataframe with 2 additional columns date and time, minus the timestamp column, and with merged capture_label
}
\description{
Prepares data processed with TrapTagger to be standardized by 
creating columns date and time from timestamp and formatting
capture_label_xx into a unique column.
}
