% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_handling.R
\name{read_snapshot_file}
\alias{read_snapshot_file}
\title{Read a file}
\usage{
read_snapshot_file(filename, base_folder, verbose = FALSE)
}
\arguments{
\item{filename}{The name of the file which should exist in base_folder.
starting at the base of the R project.}

\item{base_folder}{The folder to look in}

\item{verbose}{logical. Display the file name?}
}
\value{
A dataframe with the contents of the csv file
}
\description{
Read a file
}
