% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{recode_behavior_digikam}
\alias{recode_behavior_digikam}
\title{Recode a behavior column}
\usage{
recode_behavior_digikam(behaviors)
}
\arguments{
\item{behavior}{a character vector of behaviors, expected to be separated by a '_&_'}
}
\value{
A dataframe with one column per behavior initially present, with
1 if the behavior was present and 0 otherwise.
}
\description{
Recode a behavior column
}
