% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{rename_standard}
\alias{rename_standard}
\title{Rename columns according to standard}
\usage{
rename_standard(
  df,
  classifier = c("zooniverse", "traptagger", "digikam"),
  standard_colnames
)
}
\arguments{
\item{df}{The dataframe with the columns to rename}

\item{classifier}{The classifier used to create the dataframe df. Can be 'zooniverse', 'traptagger', 'digikam'.}

\item{standard_colnames}{A dataframe with 2 columns (at least) named like the classifier
and 'new'.
The column named like the classifier contains column names 
that are expected in the initial file. These names will be matched in the column names of df using partial matching (case insensitive and
removing blanks).
The column 'new' contains the column names that are expected in the final file. 
Columns in the classifier column will be renamed following the name of the 
corresponding value in 'new'. If no old column corresponds to 'new' (indicated with a NA)
then the column will be created and filled with NAs.}
}
\value{
Returns a dataframe for which the columns have been renamed
}
\description{
This function renames the existing columns in a dataframe to match the standard names.
}
