% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{standard}
\alias{standard}
\title{Standard column names}
\format{
A data frame with 39 rows and 4 variables:
\describe{
\item{\code{zooniverse}}{character Expected columns names for Zooniverse data}
\item{\code{traptagger}}{character Expected columns names for TrapTagger data}
\item{\code{digikam}}{character Expected columns names for Digikam data}
\item{\code{new}}{character Columns names for the output standardized data}
}
}
\usage{
standard
}
\description{
A dataframe listing the standard column names for the different Snapshot data formats.
}
\keyword{datasets}
