% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize_columns}
\alias{standardize_columns}
\title{Standardize data}
\usage{
standardize_columns(
  df,
  classifier = c("zooniverse", "traptagger", "digikam"),
  standard_colnames,
  verbose = FALSE
)
}
\arguments{
\item{df}{The dataframe to standardize}

\item{classifier}{The classifier used to generate the data: possible values are 'zooniverse', 'traptagger', 'digikam'}

\item{standard_colnames}{A dataframe with 2 columns (at least) named like the classifier
and 'new'.
The column named like the classifier contains column names 
that are expected in the initial file. These names will be matched in the column names of df using partial matching (case insensitive and
removing blanks).
The column 'new' contains the column names that are expected in the final file. 
Columns in the classifier column will be renamed following the name of the 
corresponding value in 'new'. If no old column corresponds to 'new' (indicated with a NA)
then the column will be created and filled with NAs.}

\item{verbose}{Should the info output messages be displayed?}
}
\value{
A dataframe with 25 columns with the standardized names
}
\description{
Given a dataframe, standardizes column names to match a standard.
It removes all original columns not in the standard and adds missing columns
(which are NA)
}
