% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize_date}
\alias{standardize_date}
\title{Standardize dates}
\usage{
standardize_date(dates)
}
\arguments{
\item{dates}{A dates vector (character). Accepts all formats coercible to date
with lubridate::as_date, else will look for "\%d-\%m-\%Y" or "\%m-\%d-\%Y".}
}
\value{
A vector of dates in the format "AAAA-MM-DD".
}
\description{
Standardizes a vector of dates to the format "AAAA-MM-DD".
}
