% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize_snapshot_df}
\alias{standardize_snapshot_df}
\title{Standardize dataframe}
\usage{
standardize_snapshot_df(df, standard_df, locationID_digikam, classifier)
}
\arguments{
\item{df}{The dataframe to standardize. It is expected to match the data format for either
Zooniverse, TrapTagger or Digikam processed data (i.e. have column names defined in standfard_df).}

\item{standard_df}{The standard dataframe to match column names to the new standard.
A dataframe with >= 2 columns, one of which must be named 'zooniverse', 'digikam' or 'traptagger' and another one
must be named 'new'.}

\item{locationID_digikam}{Optional locationID to be used for Digikam data
(will display a warning if not provided for Digikam data.)}

\item{classifier}{Optional character for the classifier.}
}
\value{
The standardized dataframe: it has the same columns as specified in
standard_df$new, dates and times are standardized to "YYYY-MM-DD" and "HH:MM:SS",
capture info (locationID, cameraID, roll, capture and season if classifier is zooniverse)
is filled. Columns are also in the same order as provided in
standard_df$new and the rows are ordered by camera, date and time.
}
\description{
Standardizes a dataframe to the Snapshot standard.
}
