% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize_snapshot_list}
\alias{standardize_snapshot_list}
\title{Standardize a list of dataframes}
\usage{
standardize_snapshot_list(df_list, standard_df, classifier)
}
\arguments{
\item{df_list}{a list of dataframes.}

\item{standard_df}{The standard dataframe to match column names to the new standard.
A dataframe with >= 2 columns, one of which must be named 'zooniverse', 'digikam' or 'traptagger' and another one
must be named 'new'.}

\item{classifier}{Optional character or vector of characters for the classifier.}
}
\value{
The list of standardized dataframes: each dataframe has the same columns as specified in
standard_df$new, dates and times are standardized to "YYYY-MM-DD" and "HH:MM:SS",
capture info (locationID, cameraID, roll, capture and season if classifier is zooniverse)
is filled. Columns are also in the same order as provided in
standard_df$new and the rows are ordered by camera, date and time.
}
\description{
Standardizes a list of dataframes to the Snapshot standard.
}
