% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{standardize_species}
\alias{standardize_species}
\title{Standardize species}
\usage{
standardize_species(species)
}
\arguments{
\item{species}{vector of species names}
}
\value{
the vector of species names with names stabdardized
}
\description{
Eliminate species duplicate names (things like 'birdofprey' ans 'birdsofprey')
}
