% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize_time}
\alias{standardize_time}
\title{Standardize times}
\usage{
standardize_time(times)
}
\arguments{
\item{times}{A times vector (character). Will work if the times are
already "HH:MM:SS" or "HH:MM:SS AM/PM"}
}
\value{
The vector standardized in the format "HH:MM:SS"
}
\description{
Standardizes a times vector to match the format "HH:MM:SS"
}
