% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traptagger}
\alias{traptagger}
\title{TrapTagger sample data}
\format{
A data frame with 100 rows and 9 variables:
\describe{
\item{\code{Capture_ID}}{character ID for the capture composed of locationcapture#roll#Cam.Site}
\item{\code{Cam.Site}}{character Camera}
\item{\code{id}}{integer Picture ID}
\item{\code{latitude}}{double Camera latitude (mock data)}
\item{\code{longitude}}{double Camera longitude (mock data)}
\item{\code{timestamp}}{character Date time}
\item{\code{capture_labels}}{character Species seen on the picture}
\item{\code{capture_sighting_count}}{integer Species count}
\item{\code{capture_url}}{character Picture url (mock data)}
}
}
\usage{
traptagger
}
\description{
A dataset mimicking typical TrapTagger data (randomized rows)
}
\keyword{datasets}
