% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_handling.R
\name{write_standardized_df}
\alias{write_standardized_df}
\title{Write the standardized file}
\usage{
write_standardized_df(
  df,
  to,
  filename,
  write = TRUE,
  return_path = ifelse(write, FALSE, TRUE),
  verbose = TRUE
)
}
\arguments{
\item{df}{The standardized file}

\item{to}{The target folder to copy data in. If it does not exist, will be created.}

\item{filename}{The name to give to the file.}

\item{write}{if TRUE (default), will write the \code{df} in the
\code{to} folder, and \code{df} will be named \code{filename}.}

\item{return_path}{Should the path be returned?}

\item{verbose}{Should messages be displayed when creating a folder/file?}
}
\value{
Writes the file to the folder \code{to/filename}.
Also returns the path \code{to/filename} if \code{return_path == TRUE}.
}
\description{
Writes a file to a given location. If \code{to} does not exist, it is
created, and if \code{filename} is not provided, a default standardized name is chosen.
}
