% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_handling.R
\name{write_standardized_file}
\alias{write_standardized_file}
\title{Write the standardized file}
\usage{
write_standardized_file(df, in_filename, to)
}
\arguments{
\item{df}{The standardized file}

\item{in_filename}{The path to the original file. If it has subfolders,
the subfolder structure is copied in to.}

\item{to}{The target folder to copy data in. It must exist.}
}
\value{
Writes the file to the folder to/xxx where xxx is
the subdirectory in which the original file was in.
Also returns the path to the file.
}
\description{
Write the standardized file
}
