% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_handling.R
\name{write_standardized_list}
\alias{write_standardized_list}
\title{Write the standardized files}
\usage{
write_standardized_list(
  df_list,
  filenames,
  to,
  write = TRUE,
  return_path = ifelse(write, FALSE, TRUE),
  verbose = TRUE
)
}
\arguments{
\item{df_list}{The standardized files list to write. If the list is named,
the subdirectory structure in the names will be used to replicate the subdirectiry structure
in the destination.}

\item{filenames}{(Optional) vector of customized file names.}

\item{to}{Destination folder to write to.}

\item{write}{Should the result be written or only the path returned?}

\item{return_path}{Should the path be returned?}

\item{verbose}{Should messages be displayed when creating a folder/file?}
}
\description{
Writes a lsit of files to a given location. If \code{to} does not exist, it is
created, and if \code{filename} is not provided, a default standardized name is chosen.
}
