% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{create_logger}
\alias{create_logger}
\title{Create logger}
\usage{
create_logger(
  my_logfile,
  threshold = c("INFO", "DEBUG", "WARNING", "ERROR", "FATAL")
)
}
\arguments{
\item{my_logfile}{path to the log file (character).}

\item{threshold}{logging levels to write to the logger (see \code{loglevel} documentation in \code{log4r}).
Defalts to \code{INFO}.}
}
\value{
Returns a logger and creates a logfile at the given path. If the path given does not
exist, also creates this path.
}
\description{
Initializes a logger with a given logfile.
}
\examples{
logger <- create_logger(tempfile())
}
