% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_handling.R
\name{get_csv_files_and_folders}
\alias{get_csv_files_and_folders}
\title{Get files and folder}
\usage{
get_csv_files_and_folders(input, except)
}
\arguments{
\item{input}{a character vector of valid paths: can be files or folders, or a mix of both}

\item{except}{files to ignore (optional)}
}
\value{
A dataframe with columns folder and files, where folder is the 
folder up to a given file and file is the file.
}
\description{
Get all csv files from the input character vector. 
If 'input' is a folder, will list all files within input; if 'input'
is a file, will only list this file.
}
