% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_handling.R
\name{get_final_filename}
\alias{get_final_filename}
\title{Get final filemane}
\usage{
get_final_filename(df)
}
\arguments{
\item{df}{The dataframe to be copied. Must have columns locationID, season, roll.}
}
\value{
The filename for this file in the format locationID_Sseason_Rroll.csv
It there are several locationID, seasons or rolls, they are separated by a dash
in the filename: locationID1-locationID2...
}
\description{
Return the filename from the file columns.
}
\examples{
zooniverse_std <- standardize_snapshot_df(df = zooniverse, standard = standard)
get_final_filename(zooniverse_std)
}
