% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{guess_classifier}
\alias{guess_classifier}
\title{Guess classifier}
\usage{
guess_classifier(colnames_df, logger = NA)
}
\arguments{
\item{colnames_df}{A character vector of column names.}

\item{logger}{a \code{log4r} \code{logger} object if you want logging (can be created with \code{create_logger}),
else \code{NA}.}
}
\value{
The classifier: either \code{zooniverse}, \code{digikam} or \code{traptagger}.
}
\description{
Guesses the classifier used to annotate the data based on the column names
given in \code{colnames_df}.
}
\examples{
guess_classifier(colnames(zooniverse))
guess_classifier(colnames(traptagger))
guess_classifier(colnames(digikam))
}
