% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize_snapshot_df}
\alias{standardize_snapshot_df}
\title{Standardize dataframe}
\usage{
standardize_snapshot_df(df, standard_df, locationID_digikam, classifier)
}
\arguments{
\item{df}{The dataframe to standardize. It is expected to match the data format for either
Zooniverse, TrapTagger or Digikam processed data (i.e. have column names defined in standard_df).}

\item{standard_df}{The standard dataframe to match column names to the new standard.
A dataframe with >= 2 columns, one of which must be named \code{zooniverse}, \code{digikam} or \code{traptagger}
and another one must be named \code{new}.}

\item{locationID_digikam}{Optional character \code{locationID} to use for Digikam data
(will display a warning if not provided for Digikam data.) Indeed, for Digikam data,
the \code{locationID} cannot be inferred from other columns.}

\item{classifier}{Optional character for the classifier.}
}
\value{
The standardized dataframe: it has the same columns as specified in
\code{standard_df$new}, dates and times are standardized and some columns regarding
information on the capture are filled. Species names (\code{snapshotName}) and \code{cameraID} and \code{locationID}
are standardized (see details).
Columns are in the same order as provided in \code{standard_df$new}
and the rows are ordered by ascending camera, date and time.
}
\description{
Standardizes a dataframe to the Snapshot standard.
}
\details{
Dates and times are standardized to \code{YYYY-MM-DD} and \code{HH:MM:SS}.
The columns \code{locationID}, \code{cameraID}, \code{roll}, \code{capture}
and \code{season} (if classifier is \code{zooniverse}) are filled with meaningful information.
Species names (\code{snapshotName}) are standardized to match the existing, known species names.
For column \code{locationID}:
\itemize{
\item The location code \code{DHP} is replaced with \code{OVE} if the corresponding camera code starts with 'O'.
\item The location code \code{KGA} is replaced with \code{KHO} if the corresponding camera code starts with 'KHO'.
}

For column \code{cameraID}:
\itemize{
\item For TrapTagger data: will remove the leading location code part for all data
(eg if \code{location} is \code{ATH}, will change \code{cameras} \code{ATH_A01} -> \code{A01}).
Also, if the location code is \code{KHO}, \code{SAM} or \code{TSW}: will remove the dash in
the camera name (e.g \code{KHO_E_A01} -> \code{EA01})
\item For Zooniverse data: if the location code is \code{KHO}, will replace
\code{KHOG} with \code{E} and  \code{KHOL} with \code{M} in \code{cameras}.
If the location code is \code{DHP}, will remove leading \code{D} in \code{cameraID}.
If the location code is \code{OVE}, will remove leading \code{O} in \code{cameraID}.
\item For column \code{eventID}: the event ID formatted as season#cam_site#roll#event_no.
}
}
\examples{
standardize_snapshot_df(zooniverse, standard)
standardize_snapshot_df(traptagger, standard)
standardize_snapshot_df(digikam, standard, locationID_digikam = "MOK")
}
