% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardizeSnapshot-package.R
\docType{package}
\name{standardizeSnapshot-package}
\alias{standardizeSnapshot}
\alias{standardizeSnapshot-package}
\title{standardizeSnapshot: Standardize and clean Snapshot data}
\description{
standardizeSnapshot is a R package to standardize camera trap records files from the Snapshot Safari project. Snapshot Safari data comes in different standards, following the method that was used to classify pictures (Zooniverse, TrapTagger or Digikam). This package allows to standardize all data sources to a unique file format and then cleans the files to homogenize records.
}
\author{
\strong{Maintainer}: Lisa Nicvert \email{lisa.nicvert@univ-lyon1.fr}

}
\keyword{internal}
