% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{write_log_message}
\alias{write_log_message}
\title{Write log message}
\usage{
write_log_message(
  message,
  logger = NA,
  level = c("info", "warn", "error", "debug")
)
}
\arguments{
\item{message}{The message to display/write to the logger}

\item{logger}{Logger to write to (\code{log4r} object of class \code{logger}) (defaults to \code{NA})}

\item{level}{Logging level: either \code{info}, \code{warn}, \code{debug} or \code{error}.}
}
\value{
Will either write the message specified in \code{message} to the logger or
do nothing if logger is \code{NA}.
}
\description{
Writes a log message. If a logger is provided, writes to that logger;
if it is NA, displays a message.
}
\details{
\itemize{
\item if \code{logger} is not \code{NA}, writes a message to the logger file with the specified level.
\item else, doesn't do anything.
}
}
\examples{
logger <- create_logger(tempfile()) # Create a logger
write_log_message("Test", level = "debug", logger = logger)
write_log_message("Test", level = "info", logger = logger)
write_log_message("Test", level = "warn", logger = logger) 
write_log_message("Test", level = "error", logger = logger) 
}
