% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_handling.R
\name{write_log_message}
\alias{write_log_message}
\title{Write log message}
\usage{
write_log_message(message, logger = NA, level = "info")
}
\arguments{
\item{message}{The message to display/write to the logger}

\item{logger}{Logger to write to (log4r object of class "logger") (defaults to NA)}

\item{level}{Logging level: either 'info', 'warn', 'debug' or 'error'.}
}
\value{
Either a message or writes a log (with the logger parameters)
}
\description{
Writes a log message. If a logger is provided, writes to that logger;
if it is NA, displays a message.
}
