% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_handling.R
\name{write_standardized_list}
\alias{write_standardized_list}
\title{Write the standardized files}
\usage{
write_standardized_list(
  df_list,
  filenames,
  to,
  write = TRUE,
  return_path = ifelse(write, FALSE, TRUE),
  verbose = TRUE,
  logger = NA
)
}
\arguments{
\item{df_list}{The standardized files list to write. If the list is named,
the subdirectory structure in the names will be used to replicate the subdirectiry structure
in the destination.}

\item{filenames}{(Optional) vector of customized file names.}

\item{to}{Destination folder to write to.}

\item{write}{Should the result be written or only the path returned?}

\item{return_path}{Should the path be returned?}

\item{verbose}{Should messages be displayed when creating a folder/file?}

\item{logger}{a \code{log4r} \code{logger} object if you want logging (can be created with \code{create_logger}),
else \code{NA}.}
}
\value{
Writes the files \code{to/filename1}, \code{to/filename2}...
Also returns the paths \code{to/filename1}, \code{to/filename2}... if \code{return_path == TRUE}.
}
\description{
Writes a list of files to a given location. If \code{to} does not exist, it is
created, and if \code{filename} is not provided, a default standardized name is chosen.
}
\examples{
# Example with a subdirectory structure (inferred from the filenames)
df_list <- list(zooniverse, digikam, traptagger)
names(df_list) <- c("APN/APN.csv", "MOK/MOK.csv", "ATH/ATH.csv")
std_list <- standardize_snapshot_list(df_list, standard)

# Don't write data
write_standardized_list(std_list, 
                        to = "/home/lnicvert/Documents/PhD/Snapshot/data/2_standardized_data", 
                        write = FALSE)
# Don't write data and use custom name
write_standardized_list(std_list, 
                        to = "/home/lnicvert/Documents/PhD/Snapshot/data/2_standardized_data",
                        filenames = c("myname1.csv", "myname2.csv", "myname3.csv"),
                        write = FALSE)
# Write files to temporary location
write_standardized_list(std_list, to = tempdir())
                        
# Without a subdirectory structure (and without list names)
df_list <- list(zooniverse, digikam, traptagger)
names(df_list)[2] <- "MOK"
std_list <- standardize_snapshot_list(df_list, standard)

# Write files to temporary location
write_standardized_list(std_list, to = tempdir())
}
